+++ flash/core/data/data_source.py
@@ -160,6 +160,9 @@ class DefaultDataSources(LightningEnum):
     JSON = "json"
     DATASETS = "datasets"
     FIFTYONE = "fiftyone"
+    DATAFRAME = "data_frame"
+    LISTS = "lists"
+    SENTENCES = "sentences"
 
     # TODO: Create a FlashEnum class???
     def __hash__(self) -> int: